Fix lexer peek & backup: Allow backup independent of peeking
First of, I removed a brain fart in backup(): stepping back the current runes
width only to then reparse that rune and set the width again (which will always be
the same width!) doesn't make sense. So we can remove that.

Apart from that, we were relying on next+backup for peeking until now. As we
can only backup one call to next this meant that we could not backup after
peeking - which is unexpected and broke down the current lexing in some edge
cases the fuzzer found (in short, peeking to lex matchers consumed the backup
and would not allow backing up for lexIdent). By just doing the peeking
manually we don't touch the lexer state and can happily backup independent of
peeking.