Add support for also testing compilation errors
Until now we only tested compiled css selectors against the expected state -
which is nice in a perfect world without broken selectors. Now that we have a
fuzzer, we are getting some broken selectors and want to verify those don't
crash the parser / result in the expected error - thus we store and test the
compilation error if there is one.