Optimize selectors: Short circuit for non element nodes
As far as I can tell css selectors only target element nodes - so if we're
looking at anything other than an element node we can just return without any
further costly computation.

This makes the code a little less pretty but the benchmarks suggest that it
seems to be worth it.