Remove superfluous html.ElementNode checks
By ensuring Selectors are only called with ElementNodes we can remove quite a
few checks for that and boost our speed and a little and make the code less
verbose.

Everything that calls Match(n) on a node that hasn't been ensured to be an
ElementNode before must do so - that means First & All and all DOM walking
Selectors need to do ElementNode checks - DOM walking selectors only for nodes
other than the one passed in.

Well, at least I think this is right - don't want to write tests right now
and rather build features - we'll see if this one is right later :D