Refactor css test: Merge selector and html files
Using a separate file for the css selectors makes for more visual noise while
looking at the testdata and doesn't add any benefits (that i can see right now)
over just including the selectors in the style tag of the html.

The benefit of including the selectors in the style is that they could
potentially actually do something for visual inspection
(as we split on /{.*}/).

We also remove optional tags according to the html5 spec to clean up the test
files; html & body tags are not required most of the time - see
https://www.w3.org/TR/html5/syntax.html#optional-tags for more on this